Fix url of letters overriden with local url
